+++ controllers/userController.js
@@ -64,7 +64,7 @@ module.exports.getUserProfile=  async (req, res) => {
 // szukamy pracownika o podanym id
         const user=await User.findById(id)
 
-        if(!user){return res.status(404).json({success:false});}
+        if(!user){return res.status(404).json({success:false,data:"user is not exist"});}
         user.password=null
         res.status(200).json({success:true,data:user});
     } catch (err) {
@@ -79,7 +79,7 @@ module.exports.updateProfile= async (req, res) => {
 console.log(req.user.id.toString())
     const id=req.params.id
     console.log(id);
-    const {name, email, password,login,opis,wiek,image} = req.body;
+    const {name, email,login,opis,wiek,image} = req.body;
     try {
 
         if(!(id===req.user.id.toString()||req.user.Admin))
@@ -92,7 +92,7 @@ console.log(req.user.id.toString())
         if(!mongoose.Types.ObjectId.isValid(id))
             return  res.status(404).json({success:false,data:"podaj poprawny id"});
 
-        const user=await User.findByIdAndUpdate(id,{$set:{name,email,password,login,opis,wiek,image}})
+        const user=await User.findByIdAndUpdate(id,{$set:{name,email,login,opis,wiek,image}})
         if(!user)
             return  res.status(404).json({success:false,data:"user o podanym id nie istnieje"});
         else {
@@ -123,7 +123,8 @@ module.exports.getUserFriends= async (req, res) => {
         if(!friends)
             res.status(404).json({success:false,data:"user has no friends"});
         let friendsIds=[]
-        friends.map((friend)=>{friend.user1!==id?friendsIds.push(friend.login1):friendsIds.push(friend.login2)})
+        console.log(friends)
+        friends.map((friend)=>{friend.user1.toString()!==id?friendsIds.push(friend.login1):friendsIds.push(friend.login2)})
         res.status(200).json(friendsIds);
     } catch (err) {
         console.log(err);
@@ -138,7 +139,7 @@ module.exports.addUserFriend=  async (req, res) => {
     const { login} = req.body;
     console.log(id,login);
     const user1=req.user
-    console.log(user1.id)
+    console.log(user1._id)
     try {
 
 
@@ -149,8 +150,9 @@ module.exports.addUserFriend=  async (req, res) => {
         // const user1=await User.findById(id)
         // if(!user1)
         //     res.status(404).json({success:false,data:"user z takim id nie znalazony"})
+        // console.log(user2)
         const newfriend=new Friend({
-            user1:user1._id.toString(),
+            user1:user1.id,
             login1:user1.login,
             user2:user2._id,
             login2:user2.login
@@ -170,6 +172,10 @@ module.exports.addUserFriend=  async (req, res) => {
 
 
     } catch (err) {
+        if(err.name === 'ValidationError'){
+            const messages =Object.values(err.errors).map(value => value.message)
+            return res.status(400).json({success:false,errors:messages});
+        }
 //z
         console.log(err);
         res.status(500).json({ error: "blad podczas tworzenia friendow." });
@@ -181,6 +187,7 @@ module.exports.deleteUserFriend=  async (req, res) => {
     console.log("delete User friend")
 
     const id=req.params.id
+    // const _id=req.user._id
     const { login} = req.body;
     console.log(id,login);
 
@@ -189,12 +196,16 @@ module.exports.deleteUserFriend=  async (req, res) => {
 
 
         const user=await User.findOne({"login":login})
+        const user2=await User.findById(id)
+        if(!user2)
+            return res.status(404).json({success:false,data:"user z takim id nie znalazony "});
+
         if(!user)
             return res.status(404).json({success:false,data:"user z takim loginem nie znalazony "});
         const friend=await Friend.findOneAndDelete
-        ({$or:[{"user1":user._id,"user2":id},{"user2":id,"user1":user._id}]})
+        ({$or:[{"login1":login,"login2":user2.login},{"login1":user2.login,"login2":login}]})
         if(!friend)
-            return res.status(404).json({success:false,data:"usery pod takimi id nie sa przyjaciolmi "})
+            return res.status(404).json({success:false,data:"usery z takimi danymi nie sa przyjaciolmi "})
         else {
             res.status(200).json({success:true,data:"teraz nie przyjaciele"});
         }
